+++ kartothek/serialization/__init__.py
@@ -2,10 +2,17 @@ import pkg_resources
 
 from ._csv import CsvSerializer
 from ._generic import (
+    ConjunctionType,
     DataFrameSerializer,
+    LiteralType,
+    LiteralValue,
+    PredicatesType,
+    check_predicates,
+    columns_in_predicates,
     filter_array_like,
     filter_df,
     filter_df_from_predicates,
+    filter_predicates_by_column,
 )
 from ._parquet import ParquetSerializer
 
@@ -25,11 +32,21 @@ def default_serializer():
 
 
 __all__ = [
-    "DataFrameSerializer",
+    # Serializer classes
     "CsvSerializer",
+    "DataFrameSerializer",
     "ParquetSerializer",
     "default_serializer",
-    "filter_df",
+    # functions
+    "check_predicates",
+    "columns_in_predicates",
     "filter_array_like",
     "filter_df_from_predicates",
+    "filter_df",
+    "filter_predicates_by_column",
+    # types
+    "ConjunctionType",
+    "LiteralType",
+    "LiteralValue",
+    "PredicatesType",
 ]